fix naming ,import,nevigative conditional
